+++ Lilu/Headers/kern_patcher.hpp
@@ -291,7 +291,7 @@ public:
 	 *
 	 *  @param patch              patch to apply
 	 *  @param startingAddress    start with this address (or kext/kernel lowest address)
-	 *  @param maxSize            maximum size to look for (or kext/kernel max size)
+	 *  @param maxSize            maximum size to lookup (or kext/kernel max size)
 	 */
 	EXPORT void applyLookupPatch(const LookupPatch *patch, uint8_t *startingAddress, size_t maxSize);
 #endif /* LILU_KEXTPATCH_SUPPORT */