+++ Lilu/Headers/kern_user.hpp
@@ -483,12 +483,10 @@ private:
 	 *  @param action     passed action, we only need KAUTH_FILEOP_EXEC
 	 *  @param arg0       pointer to vnode (vnode *) for executable
 	 *  @param arg1       pointer to path (char *) to executable
-	 *  @param arg2       unused
-	 *  @param arg3       unsed
 	 *
 	 *  @return 0 to allow further execution
 	 */
-	static int execListener(kauth_cred_t credential, void *idata, kauth_action_t action, uintptr_t arg0, uintptr_t arg1, uintptr_t arg2, uintptr_t arg3);
+	static int execListener(kauth_cred_t /* credential */, void *idata, kauth_action_t action, uintptr_t /* arg0 */, uintptr_t arg1, uintptr_t, uintptr_t);
 
 	/**
 	 *  Unrestricted vm_protect, that takes care of Mojave codesign limitations for everyone's good.